fix(insights): break weekly-report/weekly-report-html circular import
weekly-report.ts and weekly-report-html.ts imported types from each other
(the html renderer imported WeeklyReportSummary etc. from weekly-report.ts,
which imports renderWeeklyReportHtml from the html file), tripping
depcruise's no-circular rule on every downstream PR. Moves the three
shared types (WeeklyTopFinding, WeeklyReportCapacity, WeeklyReportSummary)
into the existing shared types.ts, which both files already depend on.
